lsp: Refactor to organise by rst/sphinx features
Rather than organising files by LSP features (completions etc.),
organise files based on the rst/sphinx features they add support for.
This means that the `esbonio.lsp.completion` and
`esbonio.lsp.definition` modules which both had a `.domains` sub-module
have been removed.

In their place we have `esbonio.lsp.sphinx.domains` and
`esbonio.lsp.sphinx.filepaths` which hopefully makes the code easier to
maintain going forward
